Implement Phase II (get next level graph) & first Louvain working
This is the first version where the full Louvain algorithm actually
works. However, many things are left to be tested at this state and
probably not everything is working correctly yet.

Fixed a bug where params were assigned incorrectly to modularity.gain()
and where the previous_community was always assigned as best_community.