Relax "read-only" policy of graph
This has to be done so that the next level graph can be constructed
without running into read-only panics.

Also relaxed the policy for accessing "invalid nodes", i.e. nodes that
are not in the graph yet. We just return a None option or an empty
HashSet in this case and deal with it on the caller side.